Finalize configuration
