+++ package.json
@@ -19,11 +19,11 @@
     "test:cov": "jest --coverage",
     "test:debug": "node --inspect-brk -r tsconfig-paths/register -r ts-node/register node_modules/.bin/jest --runInBand",
     "test:e2e": "jest --config ./test/jest-e2e.json",
-    "typeorm": "npx ts-node ./node_modules/typeorm/cli.js --config src/config/database-config.ts -n",
-    "typeorm:create": "npx ts-node ./node_modules/typeorm/cli.js --config src/config/database-config.ts migration:create -n",
-    "typeorm:generate": "npx ts-node ./node_modules/typeorm/cli.js --config src/config/database-config.ts migration:generate -n",
-    "typeorm:revert": "npx ts-node ./node_modules/typeorm/cli.js --config src/config/database-config.ts migration:revert",
-    "typeorm-migration": "npx ts-node ./node_modules/typeorm/cli.js --config src/config/database-config.ts migration:run",
+    "typeorm": "npx ts-node ./node_modules/typeorm/cli.js --config src/config/orm.config.ts -n",
+    "typeorm:create": "npx ts-node ./node_modules/typeorm/cli.js --config src/config/orm.config.ts migration:create -n",
+    "typeorm:generate": "npx ts-node ./node_modules/typeorm/cli.js --config src/config/orm.config.ts migration:generate -n",
+    "typeorm:revert": "npx ts-node ./node_modules/typeorm/cli.js --config src/config/orm.config.ts migration:revert",
+    "typeorm-migration": "npx ts-node ./node_modules/typeorm/cli.js --config src/config/orm.config.ts migration:run",
     "create-user": "npx ts-node ./scripts/user.ts create"
   },
   "dependencies": {
