+++ sonar-project.properties
@@ -2,4 +2,4 @@ sonar.projectKey=gipfeli-io_gipfeli-api
 sonar.organization=gipfeli-io
 sonar.javascript.lcov.reportPaths=./coverage/lcov.info
 sonar.sources=src
-sonar.coverage.exclusions=**/*.spec.ts,**/database-config.ts,**/*.module.ts,**/*.entity.ts,**/*.strategy.ts,**/dto/*.ts
+sonar.coverage.exclusions=**/*.spec.ts,**/config/*.config.ts,**/*.module.ts,**/*.entity.ts,**/*.strategy.ts,**/dto/*.ts
